Move env_path and help string to functions
The env_path was being set repeatedly so I moved it to a single function
to be more dry and make it easier to refactor if needed.  The same
applies to the help sting.  The help string needs some love but that
isn't in the scope of this commit.  The env_path has been moved to make
way for a pre and post activate / deactivate feature.  Similar to
pythons virtualenvwrapper.