Add pre/post activation/deactivation hooks
The pre and post, activation and deactivation hooks give people the
ability to run scripts before and after activation or deactivation.  The
default behavior is to prepend the environment name to the front of the
prompt.  This idea was taken from the python virtualenvwrapper project.

Also i moved the tab completion function up in the file.  I think it
makes more sense organized this way.